requirements.txt
Django==2.1.7
celery==4.2.1
django-celery==3.2.2
pymysql==0.9.3
djangorestframework==3.9.2
coreapi==2.3.3
